+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The present file will list all changes made to the project; according to the
 
 ### Added
 - New schemas/endpoints for Service Catalog information in High-Level API v2.4.
+- New schemas/endpoints for reminder translations in High-Level API v2.4.
 
 ### Changed